Chrono::PE 1.0

+++ src/chrono_vehicle/ChTransmission.cpp
@@ -55,9 +55,7 @@ void ChTransmission::UpdateInertiaProperties() {}
 
 void ChTransmission::SetGear(int gear) {
     assert(gear >= -1);
-    // Note that gear can be -1 and m_gear_ratios.size() returns an unsigned value
-    // which we need to cast to int to do the right assert here.
-    assert(gear < (int) m_gear_ratios.size());
+    assert(gear < m_gear_ratios.size());
 
     if (m_current_gear == gear)
         return;
